Se implementa en backrust el Modulo Persona con sus respectivos Tipos

+++ backrust/src/api/controller/persona_controller.rs
@@ -6,7 +6,7 @@ use super::models::Persona;
 pub struct PersonaController;
 
 impl PersonaController {
-    pub fn get(id_identificacion: i32, id_persona: String) -> Result<Persona, String> {
+    pub fn get(id_identificacion: i64, id_persona: String) -> Result<Persona, String> {
         Ok(Persona::get(id_identificacion, id_persona))
     }
 }
